Add GitHub Action
Create a "matrix build" of all the different combinations of the
Kea services and Debian/Alpine releases.

This should build and tag as expected, even though the empty ""
option for the Debian OS looks a little bit weird in the Actions
tab on GitHub.

The builds are really quick (like 1 minute) so we set a quite low
timeout so we catch errors quicker.