Add Kea user + uid:gid information in container
The Kea installation process adds a specific user that owns some of
the used folders, but some are still root:root. Here I try to add
information about the username and its uid + gid which differs between
Debian and Alpine:

  Debian: _kea 101:101
  Alpine: kea 100:101

The Kea process is still started as root (since it needs those
privileges), so these permissions are irrelevant, but with this we
lay the foundation for perhaps making this run as non-root sometime in
the future.