Switch to compiling Kea
On the package mirror there only exist packages for the x86
architecture, so if we want to support Arm we need to compile
Kea by ourself.

This takes an awful long time, so we will need to experiment to see
if we can run this on GitHub Actions, or if this needs to be done
from my local workstation.