Replace GitHub Action with local shell script
When compiling Kea for the emulated systems (i.e. not x86_64) it
takes a really long time (more than 6 hours), so we need to build
these on a local computer which does not time out. I have no good
solution for this right now, but will look into alternatives in
the future.