Remove "hooks" from main image
The hooks folder is quite large compared to the rest of the image,
and the features it contains will probably not be used by the average
user so let's exclude them by default.

With this we also create another image from which advanced users may
COPY the relevant hooks from.