Add simple build test on pull requests
When a pull request is made we start a very simple build which will
just sanity check that the amd64 versions of both Debian and Apline
builds. If these fails we cannot hope that any of the other ones will
succeed.

However, this does not push anything to DockerHub since it takes more
than 6 hours for the arm/v7 stuff to build so I have to do it locally.