Change Dockerfile target names and improve build scripts
It was not uncommon that the parent containers were updated during
the time we were building which lead us to pull and do a full rebuild
on every target which would make it take days before we completed
the release process.

With this change we now only do one pull of the base image and then
we build from cache. This has the benefit of all target having the
same base and not accidentally differ.